+++ libs/shared/v2/src/lib/schemas/search/asset-search-stats.schema.ts
@@ -2,11 +2,11 @@ import { WorkflowStatus } from '@swissgeol/ui-core';
 import { Type } from 'class-transformer';
 
 import { IsArray, IsInt, IsObject } from 'class-validator';
-import { AssetSearchUsageCode } from '../../models/asset-search/asset-search-query';
-import { AssetSearchStats, ValueCount } from '../../models/asset-search/asset-search-stats';
 import { ContactId } from '../../models/contact';
 import { GeometryType } from '../../models/geometry';
 import { LocalizedItemCode } from '../../models/localized-item';
+import { AssetSearchStats, ValueCount } from '../../models/search/asset-search-stats';
+import { AssetSearchUsageCode } from '../../models/search/search-query';
 import { WorkgroupId } from '../../models/workgroup';
 import { LocalDateRangeSchema } from '../base/local-date-range.schema';
 import { Schema } from '../base/schema';
@@ -14,6 +14,9 @@ import { Schema } from '../base/schema';
 export class AssetSearchStatsSchema extends Schema implements AssetSearchStats {
   @IsInt()
   total!: number;
+
+  @IsInt()
+  totalFiles!: number;
 
   @IsArray()
   @IsObject({ each: true })